Added a QR Code Generator
The qrcodegen project was used to add a automatically generated qrcode that allows your phone to scan it to connect to the temporary Access Point. Had to pick a lower version number (5) instead of the maximum (40) due to ESP32 hardware limitations. Version 40 supports ~1800 chars (with highest ECC level), but causes some sort of stack overflow/segfault/etc on the device. Version 5 supports 154 characters (with ECC set to LOW) which should be plenty for the temporary AP Name/Password.